store should be able to be used api-less as well

+++ src/app/services/store/store.ts
@@ -1,4 +1,4 @@
-import {Inject,Injectable} from 'angular2/core';
+import {Inject,Injectable,Optional} from 'angular2/core';
 import {BehaviorSubject, Observable , Subject} from 'rxjs/Rx';
 
 import {newApiService} from '../api/api'
@@ -26,11 +26,15 @@ export class Store {
   //    store will be modelled.
   constructor(
     private _artifact:string,
-    private _capabilities:Array<string> = [ 'GET' , 'POST' , 'PUT' , 'DELETE' ],
-    @Inject( newApiService ) newApiService
+    @Optional() @Inject( newApiService ) newApiService?,
+    private _capabilities?:Array<string>
   ) {
     this._api = newApiService;
-    this.load();
+
+    if ( this._api ) {
+      this._capabilities  = [ 'GET' , 'POST' , 'PUT' , 'DELETE' ];
+      this.load();
+    }
 
     console.log( 'Init Store with' , _artifact , _capabilities , this );
   }
@@ -56,8 +60,11 @@ export class Store {
     let items = this.items$.getValue(),
         item  = { index: items.push( artifact ) , value : artifact };
 
-    return this._api.post( this._artifact , null , artifact )
-      .subscribe( res => this.items$.next( items ) );
+    if ( this._api )
+      this._api.post( this._artifact , null , artifact )
+        .subscribe( res => this.items$.next( items ) );
+
+    return item;
   }
 
   // 1. This removes an artifact of the initialized type from the store
@@ -67,20 +74,25 @@ export class Store {
     if ( ! this._can( 'DELETE' ) )
       return null;
 
-    let item  = this.find( artifact.name ),
-        items = this.items$.getValue();
+    let items = this.items$.getValue(),
+        item  = this.get( artifact );
 
     if ( item ) {
       items.splice( item.index , 1 );
-      return this._api.delete( this._artifact , item.value.name , withPayload && item.value )
+
+    if ( item && this._api )
+      this._api.delete( this._artifact , item.value.name , withPayload && item.value )
         .subscribe( res => this.items$.next( items ) );
     }
+
+    return item;
   }
 
   // Helper method to query for artifacts in the store.
   find( val , property = 'name' ) {
     let found = null,
-        items = this.items$.getValue();
+        items = this.items$.getValue(),
+        item  = null;
 
     for ( let item of items ) {
       if ( item[ property ] == val )
@@ -93,20 +105,22 @@ export class Store {
   // 1. Retrieve a single artifact from the store
   // 2. (optionally) GET's the single artifact from the API
   // Q:
-  get( id ) {
+  get( artifact ) {
     if ( ! this._can( 'GET' ) )
       return null;
 
-    let item  = this.find( id ),
-        items = this.items$.getValue();
+    // let item  = typeof artifact === 'number' ? this.items,
+    let items = this.items$.getValue(),
+        item  = typeof artifact === 'number' ? items[ artifact ] : this.find( artifact.name );
 
-    this._api.get( this._artifact , id )
-      .subscribe( res => {
-        Object.assign( items[ item.index ] , res );
-        this.items$.next( items );
-      } );
+    if ( this._api )
+      this._api.get( this._artifact , artifact.name )
+        .subscribe( res => {
+          Object.assign( item , res );
+          this.items$.next( items );
+        } );
 
-    return item && item.value;
+    return item;
   }
 
   // 1. Updates an existing artifact in the store
@@ -115,14 +129,18 @@ export class Store {
     if ( ! this._can( 'PUT' ) )
       return null;
 
-    let item  = this.find( artifact.name ),
-        items = this.items$.getValue();
+    let items = this.items$.getValue(),
+        item  = this.get( artifact );
 
     if ( item ) {
-      Object.assign( items[ item.index ] , artifact );
-      return this._api.put( this._artifact , artifact.name , artifact )
-        .subscribe( res => this.items$.next( items ) );
+      Object.assign( item , artifact );
+
+      if ( this._api )
+        this._api.put( this._artifact , artifact.name , artifact )
+          .subscribe( res => this.items$.next( items ) );
     }
+
+    return item;
   }
